+++ trilab_invoice/models/sale_order_line.py
@@ -1,6 +1,8 @@
 from odoo import models, api
 from odoo.exceptions import UserError
 
+import logging
+_logger = logging.getLogger(__name__)
 
 class SaleOrderLine(models.Model):
     _inherit = 'sale.order.line'